+++ tensorcircuit/compiler/__init__.py
@@ -4,5 +4,6 @@ only reserve for internal use
 """
 
 from .composed_compiler import Compiler, DefaultCompiler, default_compile
+from .symbolic_compiler import lightcone_compile
 from . import simple_compiler
 from . import qiskit_compiler